+++ tests/integration/CreditCardTest.php
@@ -19,6 +19,8 @@ class Braintree_CreditCardTest extends PHPUnit_Framework_TestCase
         $this->assertEquals('Cardholder', $result->creditCard->cardholderName);
         $this->assertEquals('05/2012', $result->creditCard->expirationDate);
         $this->assertEquals(1, preg_match('/\A\w{32}\z/', $result->creditCard->uniqueNumberIdentifier));
+        $this->assertFalse($result->creditCard->isVenmoSdk());
+        $this->assertEquals(1, preg_match('/png/', $result->creditCard->imageUrl));
     }
 
     function testCreate_withDefault()
@@ -220,6 +222,60 @@ class Braintree_CreditCardTest extends PHPUnit_Framework_TestCase
 
         $errors = $result->errors->forKey('creditCard')->forKey('billingAddress')->onAttribute('base');
         $this->assertEquals(Braintree_Error_Codes::ADDRESS_INCONSISTENT_COUNTRY, $errors[0]->code);
+    }
+
+    function testCreate_withVenmoSdkPaymentMethodCode()
+    {
+        $customer = Braintree_Customer::createNoValidate();
+        $result = Braintree_CreditCard::create(array(
+            'customerId' => $customer->id,
+            'venmoSdkPaymentMethodCode' => Braintree_Test_VenmoSdk::generateTestPaymentMethodCode("378734493671000")
+        ));
+        $this->assertTrue($result->success);
+        $this->assertEquals("378734", $result->creditCard->bin);
+    }
+
+    function testCreate_with_invalid_venmoSdkPaymentMethodCode()
+    {
+        $customer = Braintree_Customer::createNoValidate();
+        $result = Braintree_CreditCard::create(array(
+            'customerId' => $customer->id,
+            'venmoSdkPaymentMethodCode' => Braintree_Test_VenmoSdk::getInvalidPaymentMethodCode()
+        ));
+        $this->assertFalse($result->success);
+        $errors = $result->errors->forKey('creditCard')->onAttribute('venmoSdkPaymentMethodCode');
+        $this->assertEquals($errors[0]->code, Braintree_Error_Codes::CREDIT_CARD_INVALID_VENMO_SDK_PAYMENT_METHOD_CODE);
+
+    }
+
+    function testCreate_with_venmoSdkSession()
+    {
+        $customer = Braintree_Customer::createNoValidate();
+        $result = Braintree_CreditCard::create(array(
+            'customerId' => $customer->id,
+            'number' => '5105105105105100',
+            'expirationDate' => '05/12',
+            'options' => array(
+                'venmoSdkSession' => Braintree_Test_VenmoSdk::getTestSession()
+            )
+        ));
+        $this->assertTrue($result->success);
+        $this->assertTrue($result->creditCard->isVenmoSdk());
+    }
+
+    function testCreate_with_invalidVenmoSdkSession()
+    {
+        $customer = Braintree_Customer::createNoValidate();
+        $result = Braintree_CreditCard::create(array(
+            'customerId' => $customer->id,
+            'number' => '5105105105105100',
+            'expirationDate' => '05/12',
+            'options' => array(
+                'venmoSdkSession' => Braintree_Test_VenmoSdk::getInvalidTestSession()
+            )
+        ));
+        $this->assertTrue($result->success);
+        $this->assertFalse($result->creditCard->isVenmoSdk());
     }
 
     function testCreateNoValidate_throwsIfValidationsFail()