+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.20.0
+
+* Fixed getting custom fields with valueForHtmlField. [Thanks to Miguel Manso for the fix.](https://github.com/mumia)
+* Adds disbursement details to transactions.
+* Adds image url to transactions.
+
 ## 2.19.0
 
 * Adds channel field to transactions.